refactor(mocks): Update package naming for clean conventions
