+++ pkg/mocks/mock_services/auth.go
@@ -1,4 +1,4 @@
-package mocks
+package mock_services
 
 import (
 	"net/http"
