+++ pkg/mocks/mock_ws/reporthub.go
@@ -1,9 +1,10 @@
-package mocks
+package mock_ws
 
 import (
+	"net/http"
+
 	"github.com/kptm-tools/core-service/pkg/domain"
 	"github.com/kptm-tools/core-service/pkg/interfaces"
-	"net/http"
 )
 
 type MockReportHub struct {
@@ -17,7 +18,7 @@ type MockReportHub struct {
 }
 
 func (m *MockReportHub) Run() {
-	//TODO implement me
+	// TODO implement me
 	panic("implement me")
 }
 
@@ -26,28 +27,32 @@ func (m *MockReportHub) Serve(w http.ResponseWriter, r *http.Request) {
 		m.MockServe(w, r)
 	}
 }
+
 func (m *MockReportHub) Register(client interfaces.IClient) {
 	if m.MockRegister != nil {
 		m.MockRegister(client)
 	}
 }
+
 func (m *MockReportHub) Unregister(client interfaces.IClient) {
 	if m.MockUnregister != nil {
 		m.MockUnregister(client)
 	}
 }
+
 func (m *MockReportHub) GetRoomVulnerabilities(scanID string) []*domain.Vulnerability {
 	if m.MockGetRoomVulnerabilities != nil {
 		return m.MockGetRoomVulnerabilities(scanID)
 	}
 	return nil
-
 }
+
 func (m *MockReportHub) AddToRoom(scanID string) {
 	if m.MockAddToRoom != nil {
 		m.MockAddToRoom(scanID)
 	}
 }
+
 func (m *MockReportHub) RemoveFromRoom(scanID string) {
 	if m.MockRemoveFromRoom != nil {
 		m.MockRemoveFromRoom(scanID)
